% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{bw_detect_format}
\alias{bw_detect_format}
\title{Detect genomic signal file format}
\usage{
bw_detect_format(file)
}
\arguments{
\item{file}{Local signal file path.}
}
\value{One of \code{bigwig}, \code{wig}, or \code{bedgraph}.}
\description{Detect genomic signal file format}
