% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bwg-track.R
\name{metadata_bwg}
\alias{metadata_bwg}
\title{Return BwgTrack metadata}
\usage{
metadata_bwg(x)
}
\arguments{
\item{x}{A validated \code{BwgTrack} object.}
}
\value{
The standardized metadata list.
}
\description{
Return BwgTrack metadata
}
