% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-bwg.R
\name{print.bwToolsBenchmark}
\alias{print.bwToolsBenchmark}
\title{Print a bwTools benchmark}
\usage{
\method{print}{bwToolsBenchmark}(x, ...)
}
\arguments{
\item{x}{A \code{bwToolsBenchmark} object.}

\item{...}{Additional arguments reserved for future use.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a bwTools benchmark
}
