% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{bwTools-package}
\alias{bwTools}
\alias{bwTools-package}
\title{bwTools: Native R tools for genomic signal tracks}
\description{
Native R tools for reading and manipulating BigWig, WIG, and bedGraph
signal tracks. In-memory objects follow the \code{BwgTrack} contract used by
GeneTrackR and use 1-based closed coordinates.
}
\author{
\strong{Maintainer}: Rensc \email{rensc0718@163.com}

}
\keyword{internal}
