% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{detect_bwg_format}
\alias{detect_bwg_format}
\title{Detect genomic signal file format}
\usage{
detect_bwg_format(file)
}
\arguments{
\item{file}{Local signal file path.}
}
\value{
One of \code{bigwig}, \code{wig}, or \code{bedgraph}.
}
\description{
Detect genomic signal file format
}
