% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-bwg-track.R
\name{print.bwToolsTrack}
\alias{print.bwToolsTrack}
\title{Print a bwTools signal track}
\usage{
\method{print}{bwToolsTrack}(x, ...)
}
\arguments{
\item{x}{A \code{bwToolsTrack} object.}

\item{...}{Additional arguments reserved for future use.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a bwTools signal track
}
