% Generated by roxygen2: do not edit by hand
\name{samples_bwg}
\alias{samples_bwg}
\title{Return sample metadata}
\usage{
samples_bwg(x, sample_ids = NULL)
}
\arguments{
\item{x}{A validated \code{BwgTrack} object.}
\item{sample_ids}{Optional sample IDs. Defaults to all samples.}
}
\value{A copy of the standardized sample metadata data.table.}
\description{Return sample metadata}
