% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-bwg.R
\name{seqinfo_bwg}
\alias{seqinfo_bwg}
\title{Return chromosome metadata}
\usage{
seqinfo_bwg(x)
}
\arguments{
\item{x}{A local BigWig path or a \code{BwgTrack}-compatible object.}
}
\value{A data.table containing chromosome metadata.}
\description{Return chromosome metadata}
