% Generated by roxygen2: do not edit by hand
\name{seqinfo_bwg}
\alias{seqinfo_bwg}
\title{Return sample-specific chromosome metadata}
\usage{
seqinfo_bwg(x, sample_ids = NULL)
}
\arguments{
\item{x}{A validated \code{BwgTrack} object.}
\item{sample_ids}{Optional sample IDs. Defaults to all samples.}
}
\value{A standardized chromosome metadata data.table.}
\description{Return sample-specific chromosome metadata}
