% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-bwg.R
\name{summary_bwg}
\alias{summary_bwg}
\title{Summarize a BwgTrack object}
\usage{
summary_bwg(x)
}
\arguments{
\item{x}{A validated \code{BwgTrack} object.}
}
\value{
A one-row data.table describing the object contract and dimensions.
}
\description{
Summarize a BwgTrack object
}
