% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-bwg.R
\name{summary_bwg}
\alias{summary_bwg}
\title{Summarize a BwgTrack-compatible object}
\usage{
summary_bwg(object)
}
\arguments{
\item{object}{A \code{BwgTrack}-compatible object.}
}
\value{
A named list with object-level counts.
}
\description{
Summarize a BwgTrack-compatible object
}
